+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Cover `<ListView>` component with unit tests (UIDATIMP-709)
 * Cover `<MARCTableView>` component with unit tests (UIDATIMP-962)
 * Remove Acquisitions accordion from the holdings field mapping screen (UIDATIMP-820)
+* Cover `<OverrideProtectedFieldsTable>` component with unit tests (UIDATIMP-715)
 
 ## [4.1.4](https://github.com/folio-org/ui-data-import/tree/v4.1.4) (2021-08-23)